feat: Allow serialization/deserialization of arbitrary CAR headers (#1)
Introduces a breaking API change from iroh-car in order to allow for the serialization/deserialization of arbitrary metadata in CAR headers.

See:

- https://github.com/n0-computer/iroh-car/issues/44
- https://github.com/darobin/dasl.ing/issues/32

## Motivation

The choice of dag-cbor map for CAR headers allows for the possibility of open-ended metadata in CAR headers, similar to the role headers play in HTTP or email. The [DASL CAR spec](https://dasl.ing/car.html
) makes this metadata support explicit.

## Implementation

[iroh-car](https://crates.io/crates/iroh-car) uses [serde_ipld_dag_cbor](https://crates.io/crates/serde_ipld_dagcbor/0.6.3) to parse the CAR header, which uses [cbor4ii](https://crates.io/crates/cbor4ii/0.2.14). cbor4ii is focused on zero-copy perf and doesn't expose a general Value type modeling the CBOR data model. Instead, it focuses on allowing you to serialize/deserialize from fixed types.

Since there is no intermediate representation of the CBOR, the minimally invasive path to supporting open-ended metadata in CAR headers is saving the literal header byte string on the CarReader struct, and providing methods to deserialize headers on-demand with Serde.

This makes a few small but breaking changes to the API:

- `CarReader.header()` returns a result, since it is now deserialized on-demand
- `CarWriter` needs an additional type parameter to allow for arbitrary serializable headers

In addition, a new method `CarReader.deserialize_header()` allows deserializing arbitrary Serde data types from the header byte slice.

The library does not do any hand-holding w/r/t forcing you to serialize valid CAR header shapes. However, if you use the `CarHeader` struct as before, it will always serialize to a valid header.

### Alternatives

I looked at using cborium or dcbor, both of which expose a Value enum type, but there was a lot of yak shaving involved.